Support libc linking with the wasm32-wasi target
This works around a couple issues:
- Our main() function assumes 3 parameters, but the signature
of the WASI function doesn't have the environment parameter.
- __stack_chk_fail() must be defined. This shouldn't be necessary,
and will never be called in webassembly anyway. We may not be
passing the correct compiler flags here.
- In WebAssembly, symbols from libc are present in the "env" module,
but we link them with `extern "c"`, causing WebAssembly runtimes to
look for them in a nonexistent "c" module.

These are workarounds, not proper fixes.