std.crypto: make ghash faster for small messages
Aggregated reduction requires 5 additional multiplications (to
precompute the powers of H), in order to save 2 multiplications
per batch.

So, only use large batches when it's actually interesting to do so.

For the last blocks, reuse the precomputations in order to perform
a single reduction.

Also, even in .ReleaseSmall, allow 2-block aggregation.
The speedup is worth it, and the code increase is reasonable.

And in .ReleaseFast, bump the upper batch size to 16.

std/crypto/benchmark.zig:

Zig 0.10.0: 2769 MiB/s
    Before: 6014 MiB/s
     After: 7104 MiB/s